// src/pets/application/dtos/pets-output.dto.ts
import { User } from "@/users/infrastructure/typeorm/entities/users.entity";

export type PetsOutput = {
  id: string;
  name: string;
  age: number;
  description: string;
  weight: number;
  color: string;
  images: string[];
  available: boolean;
  owner: User;
  created_at: Date;
  updated_at: Date;
};